backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
email-validator
dnspython
